simplest possible test working and ready for my review
