+++ .github/workflows/tests.yaml
@@ -999,8 +999,8 @@ jobs:
         run: |
           cd apps/upgrade-journey-raft
           ./scripts/metadata_only_voters_check.sh
-  corrupted-tenants:
-    name: Corrupted tenants still succeed with valid quorum
+  corrupt-shards:
+    name: Test operations on shards with corrupted files
     runs-on: ubuntu-latest-8-cores
     timeout-minutes: 60
     env:
@@ -1018,4 +1018,4 @@ jobs:
           username: ${{secrets.DOCKER_USERNAME}}
           password: ${{secrets.DOCKER_PASSWORD}}
       - name: Run chaos test
-        run: ./corrupted_tenants.sh
+        run: ./corrupt_shards.sh
