--- apps/corrupted-tenants/food.go
@@ -1,642 +1,0 @@
-//                           _       _
-// __      _____  __ ___   ___  __ _| |_ ___
-// \ \ /\ / / _ \/ _` \ \ / / |/ _` | __/ _ \
-//  \ V  V /  __/ (_| |\ V /| | (_| | ||  __/
-//   \_/\_/ \___|\__,_| \_/ |_|\__,_|\__\___|
-//
-//  Copyright © 2016 - 2023 Weaviate B.V. All rights reserved.
-//
-//  CONTACT: hello@weaviate.io
-//
-
-package main
-
-import (
-	"context"
-	"fmt"
-
-	"github.com/go-openapi/strfmt"
-	"github.com/weaviate/weaviate-go-client/v4/weaviate"
-	"github.com/weaviate/weaviate/entities/models"
-	"github.com/weaviate/weaviate/entities/schema"
-)
-
-const (
-	PIZZA_QUATTRO_FORMAGGI_ID = "10523cdd-15a2-42f4-81fa-267fe92f7cd6"
-	PIZZA_FRUTTI_DI_MARE_ID   = "927dd3ac-e012-4093-8007-7799cc7e81e4"
-	PIZZA_HAWAII_ID           = "00000000-0000-0000-0000-000000000000"
-	PIZZA_DOENER_ID           = "5b6a08ba-1d46-43aa-89cc-8b070790c6f2"
-	SOUP_CHICKENSOUP_ID       = "8c156d37-81aa-4ce9-a811-621e2702b825"
-	SOUP_BEAUTIFUL_ID         = "27351361-2898-4d1a-aad7-1ca48253eb0b"
-	RISOTTO_RISI_E_BISI_ID    = "da751a25-f573-4715-a893-e607b2de0ba4"
-	RISOTTO_ALLA_PILOTA_ID    = "10c2ee44-7d58-42be-9d64-5766883ca8cb"
-	RISOTTO_AL_NERO_DI_SEPPIA = "696bf381-7f98-40a4-bcad-841780e00e0e"
-)
-
-var IdsByClass = map[string][]string{
-	"Pizza": {
-		PIZZA_QUATTRO_FORMAGGI_ID,
-		PIZZA_FRUTTI_DI_MARE_ID,
-		PIZZA_HAWAII_ID,
-		PIZZA_DOENER_ID,
-	},
-	"Soup": {
-		SOUP_CHICKENSOUP_ID,
-		SOUP_BEAUTIFUL_ID,
-	},
-	"Risotto": {
-		RISOTTO_RISI_E_BISI_ID,
-		RISOTTO_ALLA_PILOTA_ID,
-		RISOTTO_AL_NERO_DI_SEPPIA,
-	},
-}
-
-var AllIds = []string{
-	PIZZA_QUATTRO_FORMAGGI_ID,
-	PIZZA_FRUTTI_DI_MARE_ID,
-	PIZZA_HAWAII_ID,
-	PIZZA_DOENER_ID,
-
-	SOUP_CHICKENSOUP_ID,
-	SOUP_BEAUTIFUL_ID,
-
-	RISOTTO_RISI_E_BISI_ID,
-	RISOTTO_ALLA_PILOTA_ID,
-	RISOTTO_AL_NERO_DI_SEPPIA,
-}
-
-// ##### SCHEMA #####
-
-func CreateSchemaPizza(client *weaviate.Client) {
-	createSchema(client, classPizza())
-}
-
-func CreateSchemaSoup(client *weaviate.Client) {
-	createSchema(client, classSoup())
-}
-
-func CreateSchemaRisotto(client *weaviate.Client) {
-	createSchema(client, classRisotto())
-}
-
-func CreateSchemaFood(client *weaviate.Client) {
-	CreateSchemaPizza(client)
-	CreateSchemaSoup(client)
-	CreateSchemaRisotto(client)
-}
-
-func CreateSchemaPizzaForTenants(client *weaviate.Client) {
-	createSchema(client, classForTenants(classPizza()))
-}
-
-func CreateSchemaSoupForTenants(client *weaviate.Client) {
-	createSchema(client, classForTenants(classSoup()))
-}
-
-func CreateSchemaRisottoForTenants(client *weaviate.Client) {
-	createSchema(client, classForTenants(classRisotto()))
-}
-
-func CreateSchemaFoodForTenants(client *weaviate.Client) {
-	CreateSchemaPizzaForTenants(client)
-	CreateSchemaSoupForTenants(client)
-	CreateSchemaRisottoForTenants(client)
-}
-
-func createSchema(client *weaviate.Client, class *models.Class) {
-	err := client.Schema().ClassCreator().
-		WithClass(class).
-		Do(context.Background())
-
-	requireNil(err)
-}
-
-// ##### CLASSES #####
-
-func classPizza() *models.Class {
-	return &models.Class{
-		Class:       "Pizza",
-		Description: "A delicious religion like food and arguably the best export of Italy.",
-		// TODO hnsw config so i can corrupt that too
-		InvertedIndexConfig: &models.InvertedIndexConfig{IndexTimestamps: true},
-		Properties:          classPropertiesFood(),
-		ShardingConfig: map[string]interface{}{
-			"desiredCount": 1,
-		},
-		ReplicationConfig: &models.ReplicationConfig{
-			Factor: 3,
-		},
-	}
-}
-
-// TODO delete soup risotto etc
-func classSoup() *models.Class {
-	return &models.Class{
-		Class:       "Soup",
-		Description: "Mostly water based brew of sustenance for humans.",
-		Properties:  classPropertiesFood(),
-	}
-}
-
-func classRisotto() *models.Class {
-	return &models.Class{
-		Class:               "Risotto",
-		Description:         "Risotto is a northern Italian rice dish cooked with broth.",
-		InvertedIndexConfig: &models.InvertedIndexConfig{IndexTimestamps: true},
-		Properties:          classPropertiesFood(),
-	}
-}
-
-func classForTenants(class *models.Class) *models.Class {
-	class.MultiTenancyConfig = &models.MultiTenancyConfig{
-		Enabled: true,
-	}
-	return class
-}
-
-func classPropertiesFood() []*models.Property {
-	nameProperty := &models.Property{
-		Name:         "name",
-		Description:  "name",
-		DataType:     schema.DataTypeText.PropString(),
-		Tokenization: models.PropertyTokenizationField,
-	}
-	descriptionProperty := &models.Property{
-		Name:        "description",
-		Description: "description",
-		DataType:    schema.DataTypeText.PropString(),
-	}
-	bestBeforeProperty := &models.Property{
-		Name:        "best_before",
-		Description: "You better eat this food before it expires",
-		DataType:    schema.DataTypeDate.PropString(),
-	}
-	priceProperty := &models.Property{
-		Name:        "price",
-		Description: "price",
-		DataType:    schema.DataTypeNumber.PropString(),
-		ModuleConfig: map[string]interface{}{
-			"text2vec-contextionary": map[string]interface{}{
-				"skip": true,
-			},
-		},
-	}
-
-	return []*models.Property{
-		nameProperty, descriptionProperty, bestBeforeProperty, priceProperty,
-	}
-}
-
-// ##### DATA #####
-
-func CreateDataPizza(client *weaviate.Client) {
-	createData(client, []*models.Object{
-		objectPizzaQuattroFormaggi(),
-		objectPizzaFruttiDiMare(),
-		objectPizzaHawaii(),
-		objectPizzaDoener(),
-	})
-}
-
-func GetDataPizza(client *weaviate.Client, consistencyLevel string) []*models.Object {
-	resp, err := client.
-		Data().
-		ObjectsGetter().
-		WithClassName("Pizza").
-		WithConsistencyLevel(consistencyLevel).
-		WithID(PIZZA_QUATTRO_FORMAGGI_ID).
-		Do(context.TODO())
-
-	requireNil(err)
-	requireNotNil(resp)
-	numExpected := 1 //len(IdsByClass["Pizza"])
-	numActual := len(resp)
-	requireTrue(numActual == numExpected, fmt.Sprintf("expected: %d, actual: %d", numExpected, numActual))
-	for i := 0; i < len(resp); i++ {
-		s := resp[i]
-		requireNotNil(s)
-		// TODO check returned data
-		// requireTrue(*s == "SUCCESS", "*s == SUCCESS")
-	}
-	return resp
-}
-
-func CreateDataSoup(client *weaviate.Client) {
-	createData(client, []*models.Object{
-		objectSoupChicken(),
-		objectSoupBeautiful(),
-	})
-}
-
-func CreateDataRisotto(client *weaviate.Client) {
-	createData(client, []*models.Object{
-		objectRisottoRisiEBisi(),
-		objectRisottoAllaPilota(),
-		objectRisottoAlNeroDiSeppia(),
-	})
-}
-
-func CreateDataFood(client *weaviate.Client) {
-	createData(client, []*models.Object{
-		objectPizzaQuattroFormaggi(),
-		objectPizzaFruttiDiMare(),
-		objectPizzaHawaii(),
-		objectPizzaDoener(),
-
-		objectSoupChicken(),
-		objectSoupBeautiful(),
-
-		objectRisottoRisiEBisi(),
-		objectRisottoAllaPilota(),
-		objectRisottoAlNeroDiSeppia(),
-	})
-}
-
-func createData(client *weaviate.Client, objects []*models.Object) {
-	resp, err := client.Batch().ObjectsBatcher().
-		WithObjects(objects...).
-		WithConsistencyLevel("ALL"). // TODO QUORUM
-		Do(context.Background())
-
-	requireNil(err)
-	requireNotNil(resp)
-	requireTrue(len(resp) == len(objects), "expected len(resp) == len(objects)")
-	for i := 0; i < len(resp); i++ {
-		s := resp[i].Result.Status
-		requireNotNil(s)
-		requireTrue(*s == "SUCCESS", "*s == SUCCESS")
-	}
-}
-
-func CreateDataPizzaQuattroFormaggiForTenants(client *weaviate.Client, tenantNames ...string) {
-	createDataForTenants(client, tenantNames, func() []*models.Object {
-		return []*models.Object{
-			objectPizzaQuattroFormaggi(),
-		}
-	})
-}
-
-func CreateDataPizzaFruttiDiMareForTenants(client *weaviate.Client, tenantNames ...string) {
-	createDataForTenants(client, tenantNames, func() []*models.Object {
-		return []*models.Object{
-			objectPizzaFruttiDiMare(),
-		}
-	})
-}
-
-func CreateDataPizzaHawaiiForTenants(client *weaviate.Client, tenantNames ...string) {
-	createDataForTenants(client, tenantNames, func() []*models.Object {
-		return []*models.Object{
-			objectPizzaHawaii(),
-		}
-	})
-}
-
-func CreateDataPizzaDoenerForTenants(client *weaviate.Client, tenantNames ...string) {
-	createDataForTenants(client, tenantNames, func() []*models.Object {
-		return []*models.Object{
-			objectPizzaDoener(),
-		}
-	})
-}
-
-func CreateDataPizzaForTenants(client *weaviate.Client, tenantNames ...string) {
-	createDataForTenants(client, tenantNames, func() []*models.Object {
-		return []*models.Object{
-			objectPizzaQuattroFormaggi(),
-			objectPizzaFruttiDiMare(),
-			objectPizzaHawaii(),
-			objectPizzaDoener(),
-		}
-	})
-}
-
-func CreateDataPizzaForTenantsWithIds(client *weaviate.Client, getId func() strfmt.UUID, tenantNames ...string) {
-	createDataForTenants(client, tenantNames, func() []*models.Object {
-		return []*models.Object{
-			objectPizzaQuattroFormaggiWithId(getId()),
-			objectPizzaFruttiDiMareWithId(getId()),
-			objectPizzaHawaiiWithId(getId()),
-			objectPizzaDoenerWithId(getId()),
-		}
-	})
-}
-
-func CreateDataSoupChickenForTenants(client *weaviate.Client, tenantNames ...string) {
-	createDataForTenants(client, tenantNames, func() []*models.Object {
-		return []*models.Object{
-			objectSoupChicken(),
-		}
-	})
-}
-
-func CreateDataSoupBeautifulForTenants(client *weaviate.Client, tenantNames ...string) {
-	createDataForTenants(client, tenantNames, func() []*models.Object {
-		return []*models.Object{
-			objectSoupBeautiful(),
-		}
-	})
-}
-
-func CreateDataSoupForTenants(client *weaviate.Client, tenantNames ...string) {
-	createDataForTenants(client, tenantNames, func() []*models.Object {
-		return []*models.Object{
-			objectSoupChicken(),
-			objectSoupBeautiful(),
-		}
-	})
-}
-
-func CreateDataSoupForTenantsWithIds(client *weaviate.Client, getId func() strfmt.UUID, tenantNames ...string) {
-	createDataForTenants(client, tenantNames, func() []*models.Object {
-		return []*models.Object{
-			objectSoupChickenWithId(getId()),
-			objectSoupBeautifulWithId(getId()),
-		}
-	})
-}
-
-func CreateDataRisottoForTenants(client *weaviate.Client, tenantNames ...string) {
-	createDataForTenants(client, tenantNames, func() []*models.Object {
-		return []*models.Object{
-			objectRisottoRisiEBisi(),
-			objectRisottoAllaPilota(),
-			objectRisottoAlNeroDiSeppia(),
-		}
-	})
-}
-
-func CreateDataRisottoForTenantsWithIds(client *weaviate.Client, getId func() strfmt.UUID, tenantNames ...string) {
-	createDataForTenants(client, tenantNames, func() []*models.Object {
-		return []*models.Object{
-			objectRisottoRisiEBisiWithId(getId()),
-			objectRisottoAllaPilotaWithId(getId()),
-			objectRisottoAlNeroDiSeppiaWithId(getId()),
-		}
-	})
-}
-
-func CreateDataFoodForTenants(client *weaviate.Client, tenantNames ...string) {
-	createDataForTenants(client, tenantNames, func() []*models.Object {
-		return []*models.Object{
-			objectPizzaQuattroFormaggi(),
-			objectPizzaFruttiDiMare(),
-			objectPizzaHawaii(),
-			objectPizzaDoener(),
-
-			objectSoupChicken(),
-			objectSoupBeautiful(),
-
-			objectRisottoRisiEBisi(),
-			objectRisottoAllaPilota(),
-			objectRisottoAlNeroDiSeppia(),
-		}
-	})
-}
-
-func createDataForTenants(client *weaviate.Client,
-	tenantNames []string, objectsSupplier func() []*models.Object,
-) {
-	objects := []*models.Object{}
-	for _, name := range tenantNames {
-		for _, object := range objectsSupplier() {
-			object.Tenant = name
-			objects = append(objects, object)
-		}
-	}
-	createData(client, objects)
-}
-
-// ##### OBJECTS #####
-
-func objectPizzaQuattroFormaggi() *models.Object {
-	return objectPizzaQuattroFormaggiWithId(PIZZA_QUATTRO_FORMAGGI_ID)
-}
-
-func objectPizzaQuattroFormaggiWithId(id strfmt.UUID) *models.Object {
-	return &models.Object{
-		Class: "Pizza",
-		ID:    id,
-		Properties: map[string]interface{}{
-			"name":        "Quattro Formaggi",
-			"description": "Pizza quattro formaggi Italian: [ˈkwattro forˈmaddʒi] (four cheese pizza) is a variety of pizza in Italian cuisine that is topped with a combination of four kinds of cheese, usually melted together, with (rossa, red) or without (bianca, white) tomato sauce. It is popular worldwide, including in Italy,[1] and is one of the iconic items from pizzerias's menus.",
-			"price":       float32(1.1),
-			"best_before": "2022-05-03T12:04:40+02:00",
-		},
-	}
-}
-
-func objectPizzaFruttiDiMare() *models.Object {
-	return objectPizzaFruttiDiMareWithId(PIZZA_FRUTTI_DI_MARE_ID)
-}
-
-func objectPizzaFruttiDiMareWithId(id strfmt.UUID) *models.Object {
-	return &models.Object{
-		Class: "Pizza",
-		ID:    id,
-		Properties: map[string]interface{}{
-			"name":        "Frutti di Mare",
-			"description": "Frutti di Mare is an Italian type of pizza that may be served with scampi, mussels or squid. It typically lacks cheese, with the seafood being served atop a tomato sauce.",
-			"price":       float32(1.2),
-			"best_before": "2022-05-05T07:16:30+02:00",
-		},
-	}
-}
-
-func objectPizzaHawaii() *models.Object {
-	return objectPizzaHawaiiWithId(PIZZA_HAWAII_ID)
-}
-
-func objectPizzaHawaiiWithId(id strfmt.UUID) *models.Object {
-	return &models.Object{
-		Class: "Pizza",
-		// this uuid guarantees that it's the first for cursor tests (otherwise
-		// they might be flaky if the randomly generated ids are sometimes higher
-		// and sometimes lower
-		ID: id,
-		Properties: map[string]interface{}{
-			"name":        "Hawaii",
-			"description": "Universally accepted to be the best pizza ever created.",
-			"price":       float32(1.3),
-		},
-	}
-}
-
-func objectPizzaDoener() *models.Object {
-	return objectPizzaDoenerWithId(PIZZA_DOENER_ID)
-}
-
-func objectPizzaDoenerWithId(id strfmt.UUID) *models.Object {
-	return &models.Object{
-		Class: "Pizza",
-		ID:    id,
-		Properties: map[string]interface{}{
-			"name":        "Doener",
-			"description": "A innovation, some say revolution, in the pizza industry.",
-			"price":       float32(1.4),
-		},
-	}
-}
-
-func objectSoupChicken() *models.Object {
-	return objectSoupChickenWithId(SOUP_CHICKENSOUP_ID)
-}
-
-func objectSoupChickenWithId(id strfmt.UUID) *models.Object {
-	return &models.Object{
-		Class: "Soup",
-		ID:    id,
-		Properties: map[string]interface{}{
-			"name":        "ChickenSoup",
-			"description": "Used by humans when their inferior genetics are attacked by microscopic organisms.",
-			"price":       float32(2.1),
-		},
-	}
-}
-
-func objectSoupBeautiful() *models.Object {
-	return objectSoupBeautifulWithId(SOUP_BEAUTIFUL_ID)
-}
-
-func objectSoupBeautifulWithId(id strfmt.UUID) *models.Object {
-	return &models.Object{
-		Class: "Soup",
-		ID:    id,
-		Properties: map[string]interface{}{
-			"name":        "Beautiful",
-			"description": "Putting the game of letter soups to a whole new level.",
-			"price":       float32(2.2),
-		},
-	}
-}
-
-func objectRisottoRisiEBisi() *models.Object {
-	return objectRisottoRisiEBisiWithId(RISOTTO_RISI_E_BISI_ID)
-}
-
-func objectRisottoRisiEBisiWithId(id strfmt.UUID) *models.Object {
-	return &models.Object{
-		Class: "Risotto",
-		ID:    id,
-		Properties: map[string]interface{}{
-			"name":        "Risi e bisi",
-			"description": "A Veneto spring dish.",
-			"price":       float32(3.1),
-			"best_before": "2022-05-03T12:04:40+02:00",
-		},
-	}
-}
-
-func objectRisottoAllaPilota() *models.Object {
-	return objectRisottoAllaPilotaWithId(RISOTTO_ALLA_PILOTA_ID)
-}
-
-func objectRisottoAllaPilotaWithId(id strfmt.UUID) *models.Object {
-	return &models.Object{
-		Class: "Risotto",
-		ID:    id,
-		Properties: map[string]interface{}{
-			"name":        "Risotto alla pilota",
-			"description": "A specialty of Mantua, made with sausage, pork, and Parmesan cheese.",
-			"price":       float32(3.2),
-			"best_before": "2022-05-03T12:04:40+02:00",
-		},
-	}
-}
-
-func objectRisottoAlNeroDiSeppia() *models.Object {
-	return objectRisottoAlNeroDiSeppiaWithId(RISOTTO_AL_NERO_DI_SEPPIA)
-}
-
-func objectRisottoAlNeroDiSeppiaWithId(id strfmt.UUID) *models.Object {
-	return &models.Object{
-		Class: "Risotto",
-		ID:    id,
-		Properties: map[string]interface{}{
-			"name":        "Risotto al nero di seppia",
-			"description": "A specialty of the Veneto region, made with cuttlefish cooked with their ink-sacs intact, leaving the risotto black .",
-			"price":       float32(3.3),
-		},
-	}
-}
-
-// ##### TENANTS #####
-
-func CreateTenantsPizza(client *weaviate.Client, tenants ...models.Tenant) {
-	createTenants(client, "Pizza", tenants)
-}
-
-func CreateTenantsSoup(client *weaviate.Client, tenants ...models.Tenant) {
-	createTenants(client, "Soup", tenants)
-}
-
-func CreateTenantsRisotto(client *weaviate.Client, tenants ...models.Tenant) {
-	createTenants(client, "Risotto", tenants)
-}
-
-func CreateTenantsFood(client *weaviate.Client, tenants ...models.Tenant) {
-	CreateTenantsPizza(client, tenants...)
-	CreateTenantsSoup(client, tenants...)
-	CreateTenantsRisotto(client, tenants...)
-}
-
-func createTenants(client *weaviate.Client, className string, tenants []models.Tenant) {
-	err := client.Schema().TenantsCreator().
-		WithClassName(className).
-		WithTenants(tenants...).
-		Do(context.Background())
-
-	requireNil(err)
-}
-
-type Tenants []models.Tenant
-
-func (t Tenants) Names() []string {
-	names := make([]string, len(t))
-	for i := range t {
-		names[i] = t[i].Name
-	}
-	return names
-}
-
-func (t Tenants) ByName(name string) *models.Tenant {
-	for i := range t {
-		if t[i].Name == name {
-			return &t[i]
-		}
-	}
-	return nil
-}
-
-func (t Tenants) ByNames(names ...string) Tenants {
-	var tenants Tenants
-	for _, n := range names {
-		if tenant := t.ByName(n); tenant != nil {
-			tenants = append(tenants, *tenant)
-		}
-	}
-	return tenants
-}
-
-func (t Tenants) IsStatus(status string) bool {
-	for i := range t {
-		if t[i].ActivityStatus != status {
-			return false
-		}
-	}
-	return true
-}
-
-func (t Tenants) WithStatus(status string) Tenants {
-	for i := range t {
-		t[i] = models.Tenant{
-			Name:           t[i].Name,
-			ActivityStatus: status,
-		}
-	}
-	return t
-}
-
-func (t Tenants) Merge(t2 Tenants) Tenants {
-	return append(t, t2...)
-}
